More restructure of function folders
