+++ app/resources/functions/calculator/isZeroish.ts
