+++ app/resources/functions/calculator/warpwidth.ts
